restructure files so the blog is now /blog urls
