+++ gatsby-node.js
@@ -44,7 +44,7 @@ exports.createPages = async ({ graphql, actions, reporter }) => {
   posts.forEach(({ node }, index) => {
     createPage({
       path: node.fields.slug,
-      component: path.resolve(`./src/components/post-page-template.js`),
+      component: path.resolve(`./src/components/blog/post-page-template.js`),
       context: { id: node.id },
     })
   })
