+++ src/pages/blog/about.js
@@ -1,0 +1,14 @@
+import React from "react"
+import { useSiteMetadata } from "../../hooks/use-site-metadata"
+
+const About = () => {
+  const { title, description } = useSiteMetadata()
+  return (
+    <div>
+      <h1>Welcome to {title}</h1>
+      <p>About: {description}</p>
+    </div>
+  )
+}
+
+export default About
