+++ src/pages/blog/archive.js
@@ -1,14 +1,13 @@
 import * as React from "react"
 import { graphql, Link, useStaticQuery } from "gatsby"
 
-import Layout from "./layout"
+import Layout from "../../components/blog/layout"
 
-const SideBar = () => {
+const Archive = () => {
   const data = useStaticQuery(graphql`
     query {
       allMdx(
         sort: { fields: [frontmatter___date], order: DESC }
-        limit: 3
         filter: { frontmatter: { published: { eq: true } } }
       ) {
         nodes {
@@ -30,10 +29,12 @@ const SideBar = () => {
 
   return (
     <Layout>
-      <h4>Most Recent Posts:</h4>
       <div>
+        <Link to="/blog">
+          <h4 class="link home-link">&lt;&lt;&lt; Back Home</h4>
+        </Link>
         {data.allMdx.nodes.map(({ excerpt, frontmatter, fields }) => (
-          <div class="block side-bar">
+          <div class="block">
             <Link to={fields.slug}>
               <h3 class="line-thin">{frontmatter.title}</h3>
               <p class="date futura">{frontmatter.date}</p>
@@ -42,11 +43,8 @@ const SideBar = () => {
           </div>
         ))}
       </div>
-      <Link to="/archive">
-        <h4 class="link">See All Posts</h4>
-      </Link>
     </Layout>
   )
 }
 
-export default SideBar
+export default Archive
